pyproject.toml
[project]
name = "matrix-herald-bot"
version = "0.1.0"
description = "A bot to discover and announce public matrix channels"
authors = [
    {name = "Benjamin"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "matrix-nio (>=0.25.2,<0.26.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "injector (>=0.22.0,<0.23.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
